+++ src/util/clone-child.js
@@ -1,0 +1,15 @@
+var mapEC = require('../co/mapec')
+
+module.exports = function cloneChildren(targetParent, sourceChild) {
+	if (sourceChild === null) return targetParent
+	var	sourceItem = mapEC(sourceChild),
+			sourceNext = sourceChild.nextSibling
+	if (!sourceItem) {
+		targetParent.appendChild(cloneChildren(sourceChild.cloneNode(false), sourceChild.firstChild))
+	}
+	else {
+		sourceItem.clone().moveto(targetParent)
+		if (sourceItem.factory) sourceNext = sourceItem.footer.nextSibling
+	}
+	return cloneChildren(targetParent, sourceNext)
+}